test_requirements.txt
# maintain the latest version
celery==5.0.6
codecov==2.1.10
coverage==5.3
cryptography==3.3.2
ddt==1.4.2
djangorestframework==3.12.4
edx-celeryutils==1.1.0
edx-lint==5.0.0
edx-opaque-keys
edx-submissions==3.3.1
jsonfield==3.1.0
mako==1.1.5
mock==4.0.3
pdbpp==0.10.3
pep8==1.7.0
pylint-celery==0.3
pylint-django==2.3.0
pyOpenSSL==20.0.1
pytz==2021.1
pytest==6.1.1
pytest-cov==2.12.1
pytest-django==4.1.0
pytest-mock==3.3.1
six==1.16.0
tox==3.20.1

xblock-utils==2.2.0
xblock-sdk==0.3.0
